Rewrite decompressor AppendXXX methods to use refs
Motivation
----------
This is a step towards code that doesn't require pinning which can help
with GC when compression/decompression is run a lot. GC will be able to
move memory even when in the middle of a compression or decompression
run and update the ref pointers.

Modifications
-------------
Rewrite all AppendXXX implementations in SnappyDecompressor to use
`ref byte` instead of `byte*` and use Unsafe methods to perform pointer
arithmetic.

Change source parameter to CopyBlockUnaligned64/128 to be `in` instead
of `ref` to simplify usage where the source is a `ReadOnlySpan<byte>`.

Results
-------
We generally see improved performance as well, except for a small
regression in .NET 6, which appears to be JIT related. However, overall
performance is still better than it was in the previous version and
hopefully the regression is resolved when we move forward with further
conversions to `ref byte`.

BenchmarkDotNet=v0.13.4, OS=Windows 11 (10.0.22000.1455/21H2)
Intel Core i7-10850H CPU 2.70GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK=7.0.102
  [Host]                       : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  MediumRun-.NET 6.0           : .NET 6.0.13 (6.0.1322.58009), X64 RyuJIT AVX2
  MediumRun-.NET 7.0           : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  MediumRun-.NET Framework 4.8 : .NET Framework 4.8 (4.8.4515.0), X64 RyuJIT VectorSize=256

IterationCount=15  LaunchCount=2  WarmupCount=10

|  Method |                          Job |            Runtime |      Mean |    Error |   StdDev | Ratio |
|-------- |----------------------------- |------------------- |----------:|---------:|---------:|------:|
| Pointer |           MediumRun-.NET 6.0 |           .NET 6.0 | 102.31 us | 0.196 us | 0.287 us |  1.00 |
|     Ref |           MediumRun-.NET 6.0 |           .NET 6.0 | 105.67 us | 0.509 us | 0.697 us |  1.03 |
|         |                              |                    |           |          |          |       |
| Pointer |           MediumRun-.NET 7.0 |           .NET 7.0 |  96.65 us | 0.131 us | 0.193 us |  1.00 |
|     Ref |           MediumRun-.NET 7.0 |           .NET 7.0 |  91.92 us | 0.179 us | 0.262 us |  0.95 |
|         |                              |                    |           |          |          |       |
| Pointer | MediumRun-.NET Framework 4.8 | .NET Framework 4.8 | 108.61 us | 0.805 us | 1.181 us |  1.00 |
|     Ref | MediumRun-.NET Framework 4.8 | .NET Framework 4.8 | 105.75 us | 0.113 us | 0.155 us |  0.98 |