Use ref byte for compression output pointer
Motivation
----------
Trying to eliminate the use of pinning to avoid interfering with GC.

Modifications
-------------
Replace all output pointer and emit handling in the compressor with
byte ref.

Add an extra byte to the minimum allocated output buffer to ensure the
output pointer is always within the buffer.

Replace another use of MemoryPool with ArrayPool to reduce heap allocs.