Allow shorter compression output buffer lengths
Motivation
----------
The current implementation requires that the compression output buffer
have enough room for the maximum possible compressed block size, which
is very unlikely to be required.

Modifications
-------------
Check the remaining output buffer size before outputting each fragment
to the buffer.

Add a benchmark for block compression.

Results
-------
An exception is only thrown if the output buffer is insufficient for the
actual compressed size.