Improve block benchmarks (#63)
Motivation
----------
The BlockCompressHtml benchmark doesn't use the full file. The
BlockDecompresHtml benchmark is best to use a 65KB block, but it isn't
including the cost of moving the data to the output.

Modifications
-------------
Use the full HTML fill for block compression, and copy the decompressed
data to output for decompression.

Make a few more improvements to the version comparison config.