Fix stream decompression corner case failure
Motivation
----------
In cases where the stream decompression block read hits with a start
point within the middle of a tag but then the literal is longer than the
entire block available we are incorrectly handling the _lookbackPosition
when we return.

Modifications
-------------
Refactor the handling to be a bit less confusing and correct the
result. Add a test for the failure case.

Fixes #88