Don't pass around references to the scrach buffer
Motivation
----------
The scratch buffer is primarily used on cold paths (split reads during
streaming or handling the last few tags in a block). Therefore, the
value in reducing range checks, etc, by passing the reference around is
minimal. This is mostly a holdover from when we pinned the reference,
which was more important due to pinning cost.

Modifications
-------------
Stop passing the reference to the subroutines, just make a reference as
needed.