Switch NuGet to use a license expression to assist with SBOM analysis
The COPYING.txt file, which was ported from the original C++
implementation, appears to be the basic BSD-3-Clause license except
for the parts about the testdata. However, the testdata is not included
in the NuGet package so isn't relevant.

Update the NuGet copyright to match the copyright notice in COPYING.txt.

Also embed the README.md file in the package.

Fixes #92