Throw useful exception on very short output buffers
Motivation
----------
Currently, if the consumer supplies an output buffer which is too short
a useful ArgumentException is typically thrown. However, if it's so
short that the varint encoded length cannot be written an
IndexOutOfRangeException is thrown instead.

Modifications
-------------
Redesign VarIntEncoding.Write to be VarIntEncoding.TryWrite, returning
false if the length is insufficient. Use this return value to throw a
more meaningful exception.