Throw useful exception on very short output buffers (#139)
Motivation
----------
Currently, if the consumer supplies an output buffer which is too short
a useful `ArgumentException` is typically thrown. However, if it's so
short that the varint-encoded length cannot be written, an
`IndexOutOfRangeException` is thrown instead. This also means that the
`TryCompress` variant incorrectly throws an exception rather than
returning `false`.

Modifications
-------------
Redesign `VarIntEncoding.Write` to be `VarIntEncoding.TryWrite`,
returning false if the length is insufficient. Use this return value to
return `false` from the `TryCompress` variant. This results in
`Compress` overloads throwing a more meaningful exception if `false` is
received.